requirements.txt
scikit-learn==1.3.0
pandas==2.0.3
google-cloud-bigquery==3.11.4
google-cloud-storage==2.10.0
numpy>=1.21.0
joblib>=1.2.0